+++ cliquet/tests/resource/test_views.py
@@ -260,6 +260,30 @@ class InvalidRecordTest(BaseWebTest):
                           headers=self.headers,
                           status=400)
 
+    def test_id_is_validated_on_post(self):
+        record = MINIMALIST_RECORD.copy()
+        record['id'] = 3.14
+        self.app.post_json(self.collection_url,
+                           {'data': record},
+                           headers=self.headers,
+                           status=400)
+
+    def test_id_is_preserved_on_post(self):
+        record = MINIMALIST_RECORD.copy()
+        record_id = record['id'] = '472be9ec-26fe-461b-8282-9c4e4b207ab3'
+        resp = self.app.post_json(self.collection_url,
+                                  {'data': record},
+                                  headers=self.headers)
+        self.assertEqual(resp.json['data']['id'], record_id)
+
+    def test_200_is_returned_if_id_matches_existing_record(self):
+        record = MINIMALIST_RECORD.copy()
+        record['id'] = self.record['id']
+        self.app.post_json(self.collection_url,
+                           {'data': record},
+                           headers=self.headers,
+                           status=200)
+
 
 class IgnoredFieldsTest(BaseWebTest):
     def setUp(self):
@@ -269,15 +293,6 @@ class IgnoredFieldsTest(BaseWebTest):
                                   body,
                                   headers=self.headers)
         self.record = resp.json['data']
-
-    def test_id_is_not_validated_and_overwritten(self):
-        record = MINIMALIST_RECORD.copy()
-        record['id'] = 3.14
-        body = {'data': record}
-        resp = self.app.post_json(self.collection_url,
-                                  body,
-                                  headers=self.headers)
-        self.assertNotEqual(resp.json['data']['id'], 3.14)
 
     def test_last_modified_is_not_validated_and_overwritten(self):
         record = MINIMALIST_RECORD.copy()